// FrontEnd/src/app/services/api-backend/model/rolesUsuario.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';
import { Usuario } from './usuario';


export interface RolesUsuario { 
    idRolUsario?: number;
    idRol?: number;
    idUsuario?: number;
    estado?: string | null;
    idRolNavigation?: Role;
    idUsuarioNavigation?: Usuario;
}